services/document/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0

# S3 / MinIO
boto3==1.34.0
botocore==1.34.0

# File validation
python-magic==0.4.27

